übergebe dem dac die auflösung

+++ code/thesisCode/main/core/sound/config.h
@@ -7,7 +7,7 @@
 #define SOUND_BUFFER_SIZE 256
 typedef int16_t sample;
 
-const uint8_t BITS_PER_SAMPLE = sizeof(sample) * 8;
+const unsigned int BITS_PER_SAMPLE = sizeof(sample) * 8;
 #define SAMPLE_MAX INT16_MAX
 #define SAMPLE_MIN INT16_MIN
 #define SAMPLE_RANGE (SAMPLE_MAX - SAMPLE_MIN)
